pyproject.toml
[tool.poetry]
name = "notifications-admin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.2"
ago = "~=0.0.95"
beautifulsoup4 = "^4.12.3"
blinker = "~=1.8"
exceptiongroup = "==1.2.2"
flask = "~=3.0"
flask-basicauth = "~=0.2"
flask-login = "^0.6"
flask-talisman = "*"
flask-wtf = "^1.2"
gunicorn = {version = "==22.0.0", extras = ["eventlet"]}
humanize = "~=4.10"
itsdangerous = "~=2.2"
jinja2 = "~=3.1"
newrelic = "*"
notifications-python-client = "==10.0.0"
pyexcel = "==0.7.0"
pyexcel-io = "==0.6.6"
pyexcel-ods3 = "==0.6.1"
pyexcel-xls = "==0.7.0"
pyexcel-xlsx = "==0.6.0"
openpyxl = "==3.0.10"
pyproj = "==3.6.1"
python-dotenv = "==1.0.1"
pytz = "^2024.1"
rtreelib = "==0.2.0"
werkzeug = "^3.0.3"
wtforms = "~=3.1"
markdown = "^3.5.2"
async-timeout = "^4.0.3"
bleach = "^6.1.0"
boto3 = "^1.34.156"
botocore = "^1.34.156"
cachetools = "^5.4.0"
cffi = "^1.16.0"
cryptography = "^43.0.0"
flask-redis = "^0.4.0"
geojson = "^3.1.0"
jmespath = "^1.0.1"
mistune = "0.8.4"
numpy = "^1.26.4"
ordered-set = "^4.1.0"
phonenumbers = "^8.13.43"
pycparser = "^2.22"
python-json-logger = "^2.0.7"
redis = "^5.0.8"
regex = "^2024.7.24"
s3transfer = "^0.10.2"
shapely = "^2.0.5"
smartypants = "^2.0.1"
certifi = "^2024.7.4"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
idna = "^3.7"
markupsafe = "^2.1.5"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.1"
requests = "^2.32.3"
six = "^1.16.0"
urllib3 = "^2.2.2"
webencodings = "^0.5.1"


[tool.poetry.group.dev.dependencies]
bandit = "*"
black = "^24.2.0"
coverage = "*"
freezegun = "^1.5.1"
flake8 = "^7.1.0"
flake8-bugbear = "^24.1.17"
flake8-print = "^5.0.0"
flake8-pytest-style = "^2.0.0"
isort = "^5.13.2"
jinja2-cli = {version = "==0.8.2", extras = ["yaml"]}
moto = "*"
pip-audit = "*"
pre-commit = "^3.8.0"
pytest = "^8.3.2"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
pytest-playwright = "^0.5.1"
pytest-xdist = "^3.5.0"
radon = "^6.0.1"
requests-mock = "^1.11.0"
vulture = "^2.11"
detect-secrets = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
